Create Encryption.cpp
First, a superincreasing sequence w is created. This is the basis for a private key. From this, calculate the sum.

Then, choose a number q that is greater than the sum.

Also, choose a number r that is in the range [1, q) and is coprime to q.

The private key consists of q, w and r.

To calculate a public key, generate the sequence β by multiplying each element in w by r mod q
The sequence β makes up the public key.

Say you want to encrypt "a". First convert "a" to binary. Then multiply each respective bit by the corresponding number in β.